Modularize hooks.ts and centralize session membership logic
hooks.ts (815 lines) split into 5 focused modules:
- hooks/types.ts: Event interfaces and shared types
- hooks/describeAction.ts: Tool action formatting (pure function)
- hooks/subagents.ts: SubagentStart/Stop lifecycle + FIFO spawn matching
- hooks/teamTools.ts: TeamCreate/Delete, Task/Message extraction, TeammateIdle
- hooks/index.ts: Orchestrator with auto-registration, activity handlers, git

state.ts membership centralization:
- Extracted getAgentsForSession() as single source of truth
- selectSession(), getStateForSession(), agentBelongsToSession() all delegate to it
- Eliminates 3x duplication of solo/team filtering logic

Zero test modifications — all 345 tests pass unchanged.
